Install yq with go install so it won't be part of the source code
When installing yq with go get the go.mod file is changing and affecting
the dependencies of the main module. If one of the other modules we are
using dependes on a different version of yq there is a collision between
the yq versions.
Therefore, we should install yq without changing the dependencies.

go install builds and installs packages in module-aware mode, ignoring the
go.mod file. This is useful for installing executables without affecting the
dependencies of the main module.